+++ contracts/interfaces/ITokenDistributor.sol
@@ -7,13 +7,13 @@ interface ITokenDistributor {
     /// @notice event for a new checkpoint
     /// @param time: the timestamp of the checkpoint
     /// @param tokens: amount of tokens checkpointed
-    event CheckpointToken(uint time, uint tokens);
+    event CheckpointToken(uint256 time, uint256 tokens);
 
     /// @notice event for a epoch that gets claimed
     /// @param user: who claimed the epoch
     /// @param epoch: the epoch number that was claimed
     /// @param tokens: the amount of tokens claimed
-    event EpochClaim(address user, uint epoch, uint tokens);
+    event EpochClaim(address user, uint256 epoch, uint256 tokens);
 
     // Errors
 
@@ -43,23 +43,20 @@ interface ITokenDistributor {
     /// @notice mapping for tokens allocated to each epoch
     /// @param epochNumber: the epoch number
     /// @return tokens for that epoch
-    function tokensPerEpoch(uint epochNumber) external view returns (uint);
+    function tokensPerEpoch(uint256 epochNumber) external view returns (uint256);
 
     /// @notice view function for calculating fees for an epoch
     /// based off staked balances from StakingRewardsV2
     /// @param to: the address the fees are being calculated for
     /// @param epochNumber: the epoch the fees are calculated for
     /// @return proportional amount of fees
-    function calculateEpochFees(
-        address to,
-        uint epochNumber
-    ) external view returns (uint256);
+    function calculateEpochFees(address to, uint256 epochNumber) external view returns (uint256);
 
     /// @notice represents the status of if a user already claimed their epoch
     /// @param to: the address being checked
     /// @param epochNumber: the epoch being checked
     /// @return true if the epoch has been claimed, false otherwise
-    function claimedEpoch(address to, uint epochNumber) external view returns (bool);
+    function claimedEpoch(address to, uint256 epochNumber) external view returns (bool);
 
     // Mutative Functions
 
@@ -69,10 +66,10 @@ interface ITokenDistributor {
     /// @notice claim tokens for a certain epoch
     /// @param to: address that epoch is being claimed for
     /// @param epochNumber: epoch that is being claimed
-    function claimEpoch(address to, uint epochNumber) external;
+    function claimEpoch(address to, uint256 epochNumber) external;
 
     /// @notice claim tokens for many epochs at once
     /// @param to: address that epoch is being claimed for
     /// @param epochs: all the epochs being claimed
-    function claimMany(address to, uint[] memory epochs) external;
+    function claimMany(address to, uint256[] memory epochs) external;
 }
