+++ test/foundry/utils/TokenDistributorInternals.sol
@@ -4,17 +4,15 @@ pragma solidity ^0.8.19;
 import {TokenDistributor} from "../../../contracts/TokenDistributor.sol";
 
 contract TokenDistributorInternals is TokenDistributor {
-    constructor(
-        address _kwenta,
-        address _stakingRewardsV2,
-        uint _offset
-    ) TokenDistributor(_kwenta, _stakingRewardsV2, _offset) {}
+    constructor(address _kwenta, address _stakingRewardsV2, uint256 _offset)
+        TokenDistributor(_kwenta, _stakingRewardsV2, _offset)
+    {}
 
-    function startOfWeek(uint timestamp) public view returns (uint) {
+    function startOfWeek(uint256 timestamp) public view returns (uint256) {
         return _startOfWeek(timestamp);
     }
 
-    function startOfEpoch(uint epochNumber) public view returns (uint) {
+    function startOfEpoch(uint256 epochNumber) public view returns (uint256) {
         return _startOfEpoch(epochNumber);
     }
 
@@ -22,11 +20,11 @@ contract TokenDistributorInternals is TokenDistributor {
         _checkpointWhenReady();
     }
 
-    function isEpochReady(uint epochNumber) external view {
+    function isEpochReady(uint256 epochNumber) external view {
         _isEpochReady(epochNumber);
     }
 
-    function epochFromTimestamp(uint timestamp) external view returns (uint) {
+    function epochFromTimestamp(uint256 timestamp) external view returns (uint256) {
         return _epochFromTimestamp(timestamp);
     }
 }
